Extract admin-only CountryType
